fix(view-08): consolidate Playwright spec + repair pre-existing PHPUnit asserts
Three CI fixes for PR #428:

1. settings-menu.spec.ts removed — its assertions are folded into
   modeling-shell.spec.ts (lessons.md §10: do not add new login-using
   specs once the cumulative count reaches the 5/IP/15min limit).
   The VIEW-08 surface now rides under the same login as the modeling
   shell + products list smoke.

2. AttributeTypeTest::twelveCases — bumped to 13 to match WYSIWYG
   added by VIEW-07.2 (#423). Pre-existing main failure.

3. DemoCatalogSeederTest — attribute count 27→28 + junctions 22→23
   to match the VIEW-07.2 `description_rich` wysiwyg seed. Pre-existing
   main failure.
